Change expectations for zero-S categorical SMD to NaN (failing)
Yang and Dalton 2012' SMD should be undefined for a pooled variance
covariance matrix that is a zero matrix. MASS::ginv() gives a zero
matrix, which is a correct generalized inverse. But this gives an
SMD of zero regardless of the mean vector difference (T - C), which
is odd if (T - C) is not a zero vector, like in the following 4-category
variable case (grouped by the same variable).

       1    2    3    4
  1 2717    0    0    0
  2    0 3743    0    0
  3    0    0 1623    0
  4    0    0    0  508

The SMD should be NaN to indicate the anomalies in the data.

>