Comment on equivalence to c(NaN,6), which fails if long double is disabled
The following configuration before make/make install reproduces this problem.
./configure --disable-long-double

In most platforms this option is enabled, and the result returned is
rep(NaN,6). But on sparc-sun-solaris, it is disabled, and gives the
following error, which can be reproduced with --disable-long-double
option.

1. Failure (at test-modules-smd.R#478): decent results are returned for anomalous/difficult data -------------------
svyStdDiff("onlyOne", "race", nhanesSvy) not equal to rep(NaN, 6)
6/6 mismatches (average diff: NaN).
First 6:
 pos   x     y diff
   1 NaN 0.117  NaN
   2 NaN 1.987  NaN
   3 NaN 1.129  NaN
   4 NaN 1.997  NaN
   5 NaN 1.162  NaN
   6 NaN 1.608  NaN

TODO: If the denominator is sufficienctly close to zero by some small
tolerance value, the result should be forced to be NaN.