Explain what is returned by var_label(df) (named list)
labelled::var_label(df) will return a named list of strings and
NULL. Each element is named by the corresponding variable name. The
single string is the variable label if present. If no label exists for
the variable, NULL is held. Therefore, it is a list rather than a
vector.